+++ src/features/trip/edit/components/Hero/DestinationForm/MobileDestinationsForm.tsx
@@ -13,7 +13,7 @@ import AppDialog from '@features/ui/AppDialog';
 import AppIconButton from '@features/ui/AppIconButton';
 import useDialog from '@hooks/useDialog';
 
-import { MAX_TRIP_DESTINATIONS } from '../../../costants';
+import { MAX_TRIP_DESTINATIONS } from '../../../../costants';
 import type { DestinationsFormInput } from './DestinationsForm';
 
 interface Props {
